Add code to read and display section headers; add info structs
Info structs contain some interpreted information, such as string
values for types or bool values for flags, in addition to the
entry structs containing the raw byte data for the fields.

For section headers, this currently shows just the section types.

TODO: Get rest of fields; parse and display name and tag fields.