Add macro for building an integral value from a slice of LE bytes
This reduces code duplication in defining a similar helper function
for each integral type that has a `from_le_bytes` method. It takes
the desired output type as the first match parameter, and expects a
`u8` slice as the second match parameter and a starting index as the
third match parameter. The starting index indicates the location in
the slice of the sequence of bytes that will be converted.