Add get string names from table and add to section header data
NOTE: This version is not quite right, because a couple sections
have names that are not found. It looks like the string table allows
getting substrings by indexing partway into a larger string.

To fix this we'll have to simplify our approach: In the next commit
we'll just use a simple function that reads a null-terminated string
from a slice of bytes starting at a given offset.