Start of #derive macro to generate method to parse bytes into struct
So far what we have generates a zero-initializer for the struct. Now
we need to learn a little more about using `syn` (or something else)
to match against specific types (i.e., u8, u16, ...), in order to
generate the specific functions for converting bytes into appropriate
types for the struct fields.