Finish out `FromBytes` derive macro for parsing structs from bytes
`#[derive(FromBytes)]` now adds a `parse_from_bytes` method to a
struct whose fields are each either a fixed-length u8 array, or one
of the types u8, u16, u32, or u64 -- which covers the structs used
for holding the low-level representation of header data in the ELF
parser. This method takes in a byte slice and calls the appropriate
functions to convert segments of bytes at the appropriate offsets
into values for each struct fields.

The implementation is a bit complex, which makes it a bit hard to
debug. There are pobably best practices for procedural macros that
I'm not following here. I'll keep looking into those. But, it does
run though :)