Add GNU binutils as git submodule and make simple build script
The build script makes a build directory for binutils (deleting an
old one if it is already present) and then runs the configure and
make commands to build binutils in that directory. This is a simple
approach and probably good enough from now. Hopefully we can just
have the Rust project build script run this script as part of its
work.

NOTE: We've realized that we'll probably need to just use the
binutils build process as-is for now, and probably forever. This is
because it would be very challenging to extract out some parts of the
project, as the parts are all built together with interdependencies.
And this way we can just pull changes or pin to commits from the
binutils git repository without major changes to our build process.