Start adding Rust wrapper for libbfd
NOTE: This is probably not going to work, at least not without much
difficulty. The problem is that the function returns a pointer to
a struct which itself contains function pointers and C enum types.
We can probably treat the function pointers, since we likely don't
need to call them. (Especially not for the simple test.) We could
also just assume the enums have underlying representations as ints,
but this is not guaranteed.

People say there are Rust tools (bindgen for example) to help with
this, and maybe that is the way to go. However, we're also having
some issues with figuring out the dependencies among the C libraries
in binutils.

So. The way we will go is probably to make a C library that presents
a simplified interface to libbfd, and then like the Rust code to that
library, instead of directly to bfd. That way, it will be easier to
control arguments and return types to be things that are easier to
mirror in Rust code.

Of course we can experiment with bindgen as well. But, we may end
up just implementing most of this in C and then calling a very high
level interface from Rust.