Add very minimal start of C wrapper interface library for libbfd
So far we have a test program and are just trying to get the
dependencies all linked in using CMake and the static libraries
built by the binutils build process.

Once that is done, we will start writing a higher-level interface
that wraps the libbfd functionality that we want to use and is
easier to use from Rust's FFI. There are several hurdles to using
the libbfd interface directly (like many function pointers), that
are mentioned in the last commit message on `binutils_integration`.