Add Rust FFI for bfd wrapper fn & add test to call it
This is just a start, but it shows the approach will work. It may
be easier to add a build step that adds all the dependencies of the
bfd wrapper into a single archive, but linking them manually as we
do now works okay.

Now we can work on building out the bfd / binutils wrapper more.